+++ test_1/lib/home_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('ホーム'),
+      ),
+      body: const Center(
+        child: Text('ホーム画面'),
+      ),
+    );
+  }
+}
