+++ test_1/lib/settings_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SettingsScreen extends StatelessWidget {
+  const SettingsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('設定'),
+      ),
+      body: const Center(
+        child: Text('設定画面'),
+      ),
+    );
+  }
+}
